+++ ios/Plugin/Extensions/CapacitorVideoPlayerPlugin.Fullscreen.swift
@@ -17,15 +17,15 @@ extension CapacitorVideoPlayerPlugin {
     // swiftlint:disable function_parameter_count
     func createVideoPlayerFullscreenView(
         call: CAPPluginCall, videoUrl: URL, rate: Float,
-        subTitleUrl: URL?, subTitleLanguage: String?,
-        subTitleOptions: [String: Any]?) {
+        exitOnEnd: Bool, subTitleUrl: URL?,
+        subTitleLanguage: String?, subTitleOptions: [String: Any]?) {
         DispatchQueue.main.async { [weak self] in
             let playerId: String = self?.fsPlayerId ?? "fullscreen"
             if let fullscreenView = self?.implementation
                 .createFullscreenPlayer(
                     playerId: playerId, videoUrl: videoUrl,
-                    rate: rate, subTitleUrl: subTitleUrl,
-                    language: subTitleLanguage,
+                    rate: rate, exitOnEnd: exitOnEnd,
+                    subTitleUrl: subTitleUrl, language: subTitleLanguage,
                     options: subTitleOptions) {
                 self?.videoPlayerFullScreenView = fullscreenView
                 self?.bgPlayer = self?.videoPlayerFullScreenView?.videoPlayer.player