+++ ios/Plugin/Extensions/CapacitorVideoPlayerPlugin.Observers.swift
@@ -94,10 +94,14 @@ extension CapacitorVideoPlayerPlugin {
             NotificationCenter.default.post(name: .playerFullscreenDismiss, object: nil)
             return
         }
+        guard let exitOnEnd = info["exitOnEnd"] as? Bool else {
+            NotificationCenter.default.post(name: .playerFullscreenDismiss, object: nil)
+            return
+        }
         guard let call = self.call else { return }
         self.createVideoPlayerFullscreenView(
             call: call, videoUrl: videoUrl,
-            rate: videoRate, subTitleUrl: nil,
+            rate: videoRate, exitOnEnd: exitOnEnd, subTitleUrl: nil,
             subTitleLanguage: nil, subTitleOptions: nil)
         return
     }